Prevent NOMAD_VAR_SLUG from ending in hyphen
This commit is intended to address the situation where a branch name
longer than 63 characters is truncated in a way such that it ends with a
hyphen. See, e.g.
```
$ /deploy.shusing nomad cluster https://dev.archive.org
deploying to
https://www-offshoot-webdev-7307-adjust-banner-image-sizing-collection-.dev.archive.org
+ nomad validate -var-file=env.env project.hcl
Job validation errors:1
error occurred:
    * Task group service validation failed: 1 error occurred:
    * Service name must be valid per RFC 1123 and can contain only
alphanumeric characters or dashes and must be no longer than 63
characters:
"www-offshoot-webdev-7307-adjust-banner-image-sizing-collection-"
```